+++ ui/center.py
@@ -1,0 +1,8 @@
+from PyQt5.QtWidgets import QDesktopWidget
+
+# 화면 중앙 배치
+def center(w):
+    qr = w.frameGeometry()
+    cp = QDesktopWidget().availableGeometry().center()
+    qr.moveCenter(cp)
+    w.move(qr.topLeft())
