+++ ui/main.py
@@ -1,8 +1,23 @@
 import sys
+from center import center
+from login import login
 from signUp import sign_up
-from PyQt5.QtWidgets import QApplication, QWidget, QDesktopWidget, QPushButton, \
-    QLineEdit, QLabel, QStackedWidget
+from mode import modeChoice
+from PyQt5.QtWidgets import QApplication, QStackedWidget
 
+# 요청자 / 제공자 선택 화면
+class Mode(modeChoice):
+    def __init__(self):
+        super().__init__()
+        #self.req_size.clicked.connect(self.openReqUserClass)
+        #self.shr_size.clicked.connect(self.openShrUserClass)
+
+    #def openReqUserClass(self):
+
+    #def openShrUserClass(self):
+
+
+# 회원가입 화면
 class SignUp(sign_up):
     def __init__(self):
         super().__init__()
@@ -14,12 +29,14 @@ class SignUp(sign_up):
         widget.setGeometry(300, 300, 400, 300)
         center(widget)
 
-class Login(QWidget):
+# 로그인 화면
+class Login(login):
     # don't touch
     def __init__(self):
         super().__init__()
         self.initUI()
         # 로그인 누른 후 넘어가기
+        self.login.clicked.connect(self.openModeClass)
         self.sign_up.clicked.connect(self.openSignUpClass)
 
     def openSignUpClass(self):
@@ -28,50 +45,10 @@ class Login(QWidget):
         widget.setGeometry(300, 300, 520, 300)
         center(widget)
 
-    # code
-    def initUI(self):
-        # 로그인 버튼
-        self.login = QPushButton('Login', self)
-        self.login.resize(80, 80)
-        self.login.move(300, 90)
-
-        # 회원가입 버튼
-        self.sign_up = QPushButton('회원가입', self)
-        self.sign_up.move(150, 180)
-
-        # 아이디, 비밀번호 알리기
-        label_id = QLabel('ID', self)
-        label_id.move(20, 100)
-        font_id = label_id.font()
-        font_id.setBold(True)
-        font_id.setPointSize(20)
-
-        label_pwd = QLabel('Password', self)
-        label_pwd.move(20, 140)
-        font_pwd = label_pwd.font()
-        font_pwd.setBold(True)
-        font_pwd.setPointSize(20)
-
-        # 아이디, 비밀번호 작성
-        self.id = QLineEdit(self)
-        self.id.move(90, 100)
-        self.pwd = QLineEdit(self)
-        self.pwd.setEchoMode(QLineEdit.Password)
-        self.pwd.move(90, 140)
-
-    # 아이디, 비밀번호 창
-    def onChanged(self, text):
-        self.id.setText(text)
-        self.id.adjustSize()
-        self.pwd.setText(text)
-        self.pwd.adjustSize()
-
-# 화면 중앙 배치
-def center(w):
-    qr = w.frameGeometry()
-    cp = QDesktopWidget().availableGeometry().center()
-    qr.moveCenter(cp)
-    w.move(qr.topLeft())
+    def openModeClass(self):
+        widget.setCurrentIndex(widget.currentIndex()+2)
+        widget.setGeometry(300,300, 520, 300)
+        center(widget)
 
 # don't touch
 if __name__ == '__main__':
@@ -84,10 +61,12 @@ if __name__ == '__main__':
     #레이아웃 인스턴스 생성
     login_ly = Login()
     SignUp_ly = SignUp()
+    Mode_ly = Mode()
 
     #레이아웃 스택 추가
     widget.addWidget(login_ly)
     widget.addWidget(SignUp_ly)
+    widget.addWidget(Mode_ly)
 
     # 창 띄우기 및 중간배치
     widget.setWindowTitle('DAIG')
